+++ lading/src/generator/http.rs
@@ -156,12 +156,6 @@ impl Http {
                     NonZeroU32::new(maximum_prebuild_cache_size_bytes.get_bytes() as u32)
                         .ok_or(Error::Zero)?;
                 let block_cache = match block_cache_method {
-                    block::CacheMethod::Streaming => block::Cache::stream(
-                        config.seed,
-                        total_bytes,
-                        &block_sizes,
-                        variant.clone(),
-                    )?,
                     block::CacheMethod::Fixed => {
                         block::Cache::fixed(&mut rng, total_bytes, &block_sizes, &variant)?
                     }