+++ lading/src/generator/splunk_hec.rs
@@ -201,9 +201,6 @@ impl SplunkHec {
             NonZeroU32::new(config.maximum_prebuild_cache_size_bytes.get_bytes() as u32)
                 .ok_or(Error::Zero)?;
         let block_cache = match config.block_cache_method {
-            block::CacheMethod::Streaming => {
-                block::Cache::stream(config.seed, total_bytes, &block_sizes, payload_config)?
-            }
             block::CacheMethod::Fixed => {
                 block::Cache::fixed(&mut rng, total_bytes, &block_sizes, &payload_config)?
             }